Make Node.ID a int32 instead of a int
It is defacto limited to 32 bits by the format used in deps_log.go.
Changing from int to int32 permits better binpacking of the Node
structure.

Did not observe a significant performance improvement.

+++ deps_log_test.go
@@ -84,7 +84,7 @@ func TestDepsLogTest_WriteRead(t *testing.T) {
 	for i := 0; i < len(log1.nodes()); i++ {
 		node1 := log1.nodes()[i]
 		node2 := log2.nodes()[i]
-		if i != node1.ID {
+		if int32(i) != node1.ID {
 			t.Fatal("expected equal")
 		}
 		if node1.ID != node2.ID {
